+++ sdk/synapse/azure-analytics-synapse-artifacts/src/main/java/com/azure/analytics/synapse/artifacts/models/AzureMLUpdateResourceActivity.java
@@ -18,23 +18,21 @@ import java.util.List;
 @Fluent
 public class AzureMLUpdateResourceActivity extends ExecutionActivity {
     /*
-     * Name of the Trained Model module in the Web Service experiment to be
-     * updated. Type: string (or Expression with resultType string).
+     * Name of the Trained Model module in the Web Service experiment to be updated. Type: string (or Expression with
+     * resultType string).
      */
     @JsonProperty(value = "typeProperties.trainedModelName", required = true)
     private Object trainedModelName;
 
     /*
-     * Name of Azure Storage linked service holding the .ilearner file that
-     * will be uploaded by the update operation.
+     * Name of Azure Storage linked service holding the .ilearner file that will be uploaded by the update operation.
      */
     @JsonProperty(value = "typeProperties.trainedModelLinkedServiceName", required = true)
     private LinkedServiceReference trainedModelLinkedServiceName;
 
     /*
-     * The relative file path in trainedModelLinkedService to represent the
-     * .ilearner file that will be uploaded by the update operation.  Type:
-     * string (or Expression with resultType string).
+     * The relative file path in trainedModelLinkedService to represent the .ilearner file that will be uploaded by the
+     * update operation.  Type: string (or Expression with resultType string).
      */
     @JsonProperty(value = "typeProperties.trainedModelFilePath", required = true)
     private Object trainedModelFilePath;