+++ sdk/synapse/azure-analytics-synapse-artifacts/src/main/java/com/azure/analytics/synapse/artifacts/models/OdbcTableDataset.java
@@ -19,8 +19,7 @@ import java.util.Map;
 @Fluent
 public class OdbcTableDataset extends Dataset {
     /*
-     * The ODBC table name. Type: string (or Expression with resultType
-     * string).
+     * The ODBC table name. Type: string (or Expression with resultType string).
      */
     @JsonProperty(value = "typeProperties.tableName")
     private Object tableName;